+++ plugin/src/main/java/com/stratio/cassandra/lucene/schema/mapping/GeoShapeMapper.java
@@ -1,0 +1,131 @@
+/*
+ * Licensed to STRATIO (C) under one or more contributor license agreements.
+ * See the NOTICE file distributed with this work for additional information
+ * regarding copyright ownership.  The STRATIO (C) licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.stratio.cassandra.lucene.schema.mapping;
+
+import com.google.common.base.Objects;
+import com.spatial4j.core.context.jts.JtsSpatialContext;
+import com.spatial4j.core.shape.Shape;
+import com.stratio.cassandra.lucene.IndexException;
+import com.stratio.cassandra.lucene.util.GeospatialUtils;
+import org.apache.cassandra.db.marshal.AsciiType;
+import org.apache.cassandra.db.marshal.UTF8Type;
+import org.apache.commons.lang3.StringUtils;
+import org.apache.lucene.document.Document;
+import org.apache.lucene.index.IndexableField;
+import org.apache.lucene.search.SortField;
+import org.apache.lucene.spatial.prefix.RecursivePrefixTreeStrategy;
+import org.apache.lucene.spatial.prefix.tree.GeohashPrefixTree;
+import org.apache.lucene.spatial.prefix.tree.SpatialPrefixTree;
+
+import javax.validation.constraints.NotNull;
+
+/**
+ * A {@link Mapper} to map geographical shapes represented according to the <a href="http://en.wikipedia.org/wiki/Well-known_text">
+ * Well Known Text (WKT)</a> format.
+ *
+ * This class depends on <a href="http://www.vividsolutions.com/jts">Java Topology Suite (JTS)</a>. This library can't
+ * be distributed together with this project due to license compatibility problems, but you can add it by putting <a
+ * href="http://search.maven.org/remotecontent?filepath=com/vividsolutions/jts-core/1.14.0/jts-core-1.14.0.jar">jts-core-1.14.0.jar</a>
+ * into Cassandra lib directory.
+ *
+ * Pole wrapping is not supported.
+ *
+ * @author Andres de la Pena {@literal <adelapena@stratio.com>}
+ */
+public class GeoShapeMapper extends SingleColumnMapper<String> {
+
+    public static final JtsSpatialContext SPATIAL_CONTEXT = JtsSpatialContext.GEO;
+    public static final int DEFAULT_MAX_LEVELS = 11;
+
+    /** The name of the mapped column. */
+    public final String column;
+
+    /** The max number of levels in the tree. */
+    public final int maxLevels;
+
+    /** The spatial strategy for radial distance searches. */
+    public final RecursivePrefixTreeStrategy strategy;
+
+    /**
+     * Builds a new {@link GeoShapeMapper}.
+     *
+     * @param field The name of the field.
+     * @param column The name of the column.
+     * @param validated If the field must be validated.
+     * @param maxLevels The maximum number of levels in the tree.
+     */
+    public GeoShapeMapper(String field, String column, Boolean validated, Integer maxLevels) {
+        super(field,
+              column,
+              true,
+              false,
+              validated,
+              null,
+              String.class,
+              AsciiType.instance,
+              UTF8Type.instance);
+
+        this.column = column == null ? field : column;
+
+        if (StringUtils.isWhitespace(column)) {
+            throw new IndexException("Column must not be whitespace, but found '%s'", column);
+        }
+
+        this.maxLevels = maxLevels == null ? DEFAULT_MAX_LEVELS : maxLevels;
+        SpatialPrefixTree grid = new GeohashPrefixTree(SPATIAL_CONTEXT, this.maxLevels);
+        strategy = new RecursivePrefixTreeStrategy(grid, field);
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public void addIndexedFields(Document document, String name, String value) {
+        Shape shape = GeospatialUtils.geometryFromWKT(SPATIAL_CONTEXT, value);
+        for (IndexableField field : strategy.createIndexableFields(shape)) {
+            document.add(field);
+        }
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public void addSortedFields(Document document, String name, String value) {
+        // Nothing to do here
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public SortField sortField(String name, boolean reverse) {
+        throw new IndexException("GeoShape mapper '%s' does not support simple sorting", name);
+    }
+
+    protected String doBase(String field, @NotNull Object value) {
+        return value.toString();
+    }
+
+    /** {@inheritDoc} */
+    @Override
+    public String toString() {
+        return Objects.toStringHelper(this)
+                      .add("field", field)
+                      .add("column", column)
+                      .add("validated", validated)
+                      .add("maxLevels", maxLevels)
+                      .toString();
+    }
+
+}